pyproject.toml
[tool.poetry]
name = "sky-bank"
version = "0.1.0"
description = ""
authors = ["Oleg Urpin <LegaLesnoy@mail.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^2.2.2"
openpyxl = "^3.1.5"
requests = "^2.32.3"
python-dotenv = "^1.0.1"
xmltodict = "^0.13.0"
mypy = "^1.11.2"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"
pytest-cov = "^5.0.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
